requirements.txt
flask
huey
gunicorn
yara-python
requests
apprise